+++ Nos/Views/AppView.swift
@@ -7,28 +7,59 @@
 
 import SwiftUI
 
+// Used in the NavigationStack and added as an environmentObject so that it can be used for multiple views
+class Router: ObservableObject {
+    @Published var path = NavigationPath()
+}
+
 struct AppView: View {
-    
+
     @StateObject private var appController = AppController()
     
-    @State var path = NavigationPath()
+    @EnvironmentObject var router: Router
+    
+    /// An enumeration of the navigation destinations for AppView.
+    enum Destination: String, Hashable {
+        case home
+        case relays
+        case settings
+        
+        var label: some View {
+            switch self {
+            case .home:
+                return Text(Localized.homeFeedLinkTitle.string)
+            case .relays:
+                return Text(Localized.relaysLinkTitle.string)
+            case .settings:
+                return Text(Localized.settingsLinkTitle.string)
+            }
+        }
+    }
     
     var body: some View {
+        
         Group {
             if appController.currentState == .onboarding {
                 OnboardingView(completion: appController.completeOnboarding)
             } else {
-                NavigationStack(path: $path) {
+                NavigationStack(path: $router.path) {
                     List {
-                        NavigationLink(Localized.homeFeedLinkTitle.string) {
+                        NavigationLink(value: Destination.home) { Destination.home.label }
+                        NavigationLink(value: Destination.relays) { Destination.relays.label }
+                        NavigationLink(value: Destination.settings) { Destination.settings.label }
+                    }
+                    .navigationDestination(for: Destination.self, destination: { destination in
+                        switch destination {
+                        case .home:
                             HomeFeedView()
-                        }
-                        NavigationLink(Localized.relaysLinkTitle.string) {
+                        case .relays:
                             RelayView()
-                        }
-                        NavigationLink(Localized.settingsLinkTitle.string) {
+                        case .settings:
                             SettingsView()
                         }
+                    })
+                    .navigationDestination(for: Event.self) { note in
+                        ThreadView(note: note)
                     }
                     .navigationTitle(Localized.nos.string)
                 }
@@ -43,10 +74,12 @@ struct AppView_Previews: PreviewProvider {
     static var persistenceController = PersistenceController.preview
     static var previewContext = persistenceController.container.viewContext
     static var relayService = RelayService(persistenceController: persistenceController)
+    static var router = Router()
     
     static var previews: some View {
         AppView()
             .environment(\.managedObjectContext, previewContext)
             .environmentObject(relayService)
+            .environmentObject(router)
     }
 }