+++ Nos/Views/CardButtonStyle.swift
@@ -1,0 +1,23 @@
+//
+//  CardButtonStyle.swift
+//  Nos
+//
+//  Created by Jason Cheatham on 2/16/23.
+//
+
+import SwiftUI
+
+struct CardButtonStyle: ButtonStyle {
+    func makeBody(configuration: Configuration) -> some View {
+        configuration.label
+            .offset(y: configuration.isPressed ? 3 : 0)
+            .compositingGroup()
+            .shadow(color: .cardBorderBottom, radius: 0, x: 0, y: 4)
+            .shadow(
+                color: .cardShadowBottom,
+                radius: configuration.isPressed ? 5 : 10,
+                x: 0,
+                y: configuration.isPressed ? 1 : 4
+            )
+    }
+}